--- a/examples/Interpolation.hs
+++ b/examples/Interpolation.hs
@@ -4,6 +4,7 @@ module Interpolation where
 
 import Data.Curve.Weierstrass.BN254 (Fr)
 import Data.Pairing.BN254 (getRootOfUnity)
+import qualified Data.Vector as V
 import FFT
 import Protolude
 
@@ -13,9 +14,9 @@ k = 5
 polySize :: Int
 polySize = 2 ^ k
 
-leftCoeffs, rightCoeffs :: [Fr]
-leftCoeffs = map fromIntegral [1 .. polySize]
-rightCoeffs = map fromIntegral (reverse [1 .. polySize])
+leftCoeffs, rightCoeffs :: V.Vector Fr
+leftCoeffs  = V.generate polySize fromIntegral
+rightCoeffs = V.generate polySize (fromIntegral . (polySize -))
 
 main :: IO ()
 main = do