--- a/galois-fft.cabal
+++ b/galois-fft.cabal
@@ -41,7 +41,7 @@ library
       base            >=4.10  && <5
     , elliptic-curve  >=0.3   && <0.4
     , galois-field    >=1.0   && <2.0
-    , poly            >=0.3.2 && <0.4
+    , poly            >=0.5   && <0.6
     , protolude       >=0.2   && <0.3
     , vector          >=0.12  && <0.13
 
@@ -67,7 +67,7 @@ test-suite fft-tests
     , galois-fft
     , galois-field          >=1.0   && <2.0
     , pairing               >=1.0   && <1.1
-    , poly                  >=0.3.2 && <0.4
+    , poly                  >=0.5   && <0.6
     , protolude             >=0.2   && <0.3
     , QuickCheck            >=2.13  && <2.14
     , quickcheck-instances  >=0.3   && <0.4
@@ -100,7 +100,7 @@ benchmark fft-benchmarks
     , galois-fft
     , galois-field    >=1.0   && <2.0
     , pairing         >=1.0   && <1.1
-    , poly            >=0.3.2 && <0.4
+    , poly            >=0.5   && <0.6
     , protolude       >=0.2   && <0.3
     , vector          >=0.12  && <0.13