--- a/package.yaml
+++ b/package.yaml
@@ -16,7 +16,7 @@ dependencies:
   - base           >= 4.10  && < 5
   - protolude      >= 0.2   && < 0.3
   - vector         >= 0.12  && < 0.13
-  - poly           >= 0.3.2 && < 0.4
+  - poly           >= 0.5   && < 0.6
   - elliptic-curve >= 0.3   && < 0.4
   - galois-field   >= 1.0   && < 2.0