--- a/stack.yaml
+++ b/stack.yaml
@@ -1,8 +1,10 @@
-resolver: lts-14.7
+resolver: nightly-2020-09-28
 extra-deps:
-  - elliptic-curve-0.3.0
-  - galois-field-1.0.0
-  - poly-0.3.2.0
-  - semirings-0.5.1
-  - pairing-1.0.0
-  - arithmoi-0.8.0.0
+  - poly-0.5.0.0
+  - github: adjoint-io/galois-field
+    commit: aeec6aece4f1b3304721b19d0ece3d9b76644bad
+  - github: adjoint-io/pairing
+    commit: 9e3fd026fb43fd64917a3b267d0d8d351eba77cc
+  - github: adjoint-io/elliptic-curve
+    commit: 445e196a550e36e0f25bd4d9d6a38676b4cf2be8
+allow-newer: true