I noticed that there is a room for performance improvements in `fft`.

| Size | Before, ms  | After, ms | Speedup
| :----| ----------: | --------: | -------:
| 2^5  | 1.366       | 0.330     | 4.1x
| 2^6  | 3.393       | 0.758     | 4.5x
| 2^7  | 9.531       | 1.735     | 5.5x
| 2^8  | 22.85       | 4.038     | 5.7x
| 2^9  | 56.32       | 9.557     | 5.9x
| 2^10 | 130.6       | 21.59     | 6.0x
| 2^11 | 294.0       | 47.89     | 6.1x
| 2^12 | 658.4       | 106.9     | 6.2x
| 2^13 | 1481        | 231.4     | 6.4x
| 2^14 | 3362        | 532.2     | 6.3x

(An acceleration of speedups can probably be attributed to forcing thunks earlier)

----

I can restore `type CoeffVec f = [f]`, if it is deemed to be a part of the public API and it is desirable to keep it unchanged.