@Bodigrim 2020-12-28
I'm no longer interested to pursue it, because `type CoeffVec = Vector` breaks `arithmetic-circuits` and I do not have much time to invest into fixing it. On the other hand, keeping `type CoeffVec = []` is not satisfactory for my applications.